requirements_compatible.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
redis>=5.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
yfinance>=0.2.0
python-dateutil>=2.8.0
tabulate>=0.9.0